Unit test added to Calculation classes
Was thinking a lot on the scope, but I deliberately didn't want to
include any other aspects of our program, so I mocked the data with
wired in constants instead of using a mocked calculationSource object,
as it would incorporate the mechanism of another class into the test.
I think that shoud be tested by another Test.
It looks funky (hardwired constants), but as these classes (and any
further similar ones we might add, e.g.: sqrt) do literally nothing more
other than applying a certain math operation, so I hope this is the good
way to go about this.